+++ demo/basic/dark-theme.component.ts
@@ -1,0 +1,60 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'basic-dark-theme-demo',
+  template: `
+    <div>
+      <h3>
+        Dark Theme
+        <small>
+          <a href="https://github.com/swimlane/ngx-datatable/blob/master/demo/basic/dark-theme.component.ts" target="_blank">
+            Source
+          </a>
+        </small>
+      </h3>
+      <ngx-datatable
+        class="dark"
+        [rows]="rows"
+        [loadingIndicator]="loadingIndicator"
+        [columns]="columns"
+        [columnMode]="'force'"
+        [headerHeight]="40"
+        [footerHeight]="40"
+        [limit]="10"
+        [rowHeight]="'auto'"
+        [reorderable]="reorderable">
+      </ngx-datatable>
+    </div>
+  `
+})
+export class DarkThemeComponent {
+
+  rows = [];
+  loadingIndicator: boolean = true;
+  reorderable: boolean = true;
+
+  columns = [
+    { prop: 'name' },
+    { name: 'Gender' },
+    { name: 'Company' }
+  ];
+
+  constructor() {
+    this.fetch((data) => {
+      this.rows = data;
+      setTimeout(() => { this.loadingIndicator = false; }, 1500);
+    });
+  }
+
+  fetch(cb) {
+    const req = new XMLHttpRequest();
+    req.open('GET', `assets/data/company.json`);
+
+    req.onload = () => {
+      cb(JSON.parse(req.response));
+    };
+
+    req.send();
+  }
+
+}