+++ demo/module.ts
@@ -20,6 +20,8 @@ import { RxDemoComponent } from './basic/rx.component';
 import { ContextMenuDemoComponent } from './basic/contextmenu.component';
 import { RowCssComponent } from './basic/css.component';
 import { DynamicHeightComponent } from './basic/dynamic-height.component';
+import { DarkThemeComponent } from './basic/dark-theme.component';
+import { FooterDemoComponent } from './basic/footer.component';
 
 // -- Paging
 import { ClientPagingComponent } from './paging/paging-client.component';
@@ -77,6 +79,7 @@ import { ColumnPinningComponent } from './columns/pinning.component';
     ColumnForceComponent,
     ColumnPinningComponent,
     FilterBarComponent,
+    DarkThemeComponent,
     TabsDemoComponent,
     SingleSelectionComponent,
     LiveDataComponent,
@@ -86,7 +89,8 @@ import { ColumnPinningComponent } from './columns/pinning.component';
     CheckboxSelectionComponent,
     MultiClickSelectionComponent,
     RowCssComponent,
-    DynamicHeightComponent
+    DynamicHeightComponent,
+    FooterDemoComponent
   ],
   imports: [BrowserModule, NgxDatatableModule],
   bootstrap: [AppComponent]