+++ release/components/footer/footer.directive.d.ts
@@ -1,0 +1,11 @@
+import { TemplateRef } from '@angular/core';
+export declare class DatatableFooterDirective {
+    footerHeight: number;
+    totalMessage: string;
+    selectedMessage: string | boolean;
+    pagerLeftArrowIcon: string;
+    pagerRightArrowIcon: string;
+    pagerPreviousIcon: string;
+    pagerNextIcon: string;
+    template: TemplateRef<any>;
+}