+++ release/datatable.module.js
@@ -20,6 +20,7 @@ NgxDatatableModule.decorators = [
                     services_1.ScrollbarHelper
                 ],
                 declarations: [
+                    components_1.DataTableFooterTemplateDirective,
                     directives_1.VisibilityDirective,
                     directives_1.DraggableDirective,
                     directives_1.ResizeableDirective,
@@ -41,7 +42,8 @@ NgxDatatableModule.decorators = [
                     components_1.DataTableBodyCellComponent,
                     components_1.DataTableSelectionComponent,
                     components_1.DataTableColumnHeaderDirective,
-                    components_1.DataTableColumnCellDirective
+                    components_1.DataTableColumnCellDirective,
+                    components_1.DatatableFooterDirective
                 ],
                 exports: [
                     components_1.DatatableComponent,
@@ -49,7 +51,9 @@ NgxDatatableModule.decorators = [
                     components_1.DatatableRowDetailTemplateDirective,
                     components_1.DataTableColumnDirective,
                     components_1.DataTableColumnHeaderDirective,
-                    components_1.DataTableColumnCellDirective
+                    components_1.DataTableColumnCellDirective,
+                    components_1.DataTableFooterTemplateDirective,
+                    components_1.DatatableFooterDirective
                 ]
             },] },
 ];