+++ src/components/body/body.component.ts
@@ -230,10 +230,11 @@ export class DataTableBodyComponent implements OnInit, OnDestroy {
    */
   ngOnInit(): void {
     if(this.rowDetail) {
-      this.listener = this.rowDetail.toggle.subscribe(({ type, value }) => {
-        if(type === 'row') this.toggleRowExpansion(value);
-        if(type === 'all') this.toggleAllRows(value);
-      });
+      this.listener = this.rowDetail.toggle
+        .subscribe(({ type, value }: { type: string, value: any }) => {
+          if(type === 'row') this.toggleRowExpansion(value);
+          if(type === 'all') this.toggleAllRows(value);
+        });
     }
   }
 
@@ -412,7 +413,7 @@ export class DataTableBodyComponent implements OnInit, OnDestroy {
    * 
    * @memberOf DataTableBodyComponent
    */
-  getDetailRowHeight(row?: any, index?: any): number {
+  getDetailRowHeight = (row?: any, index?: any): number => {
     if (!this.rowDetail) return 0;
     const rowHeight = this.rowDetail.rowHeight;
     return typeof rowHeight === 'function' ? rowHeight(row, index) : rowHeight;
@@ -515,7 +516,7 @@ export class DataTableBodyComponent implements OnInit, OnDestroy {
     // Initialize the tree only if there are rows inside the tree.
     if (this.rows && this.rows.length) {
       this.rowHeightsCache.initCache(
-        this.rows, this.rowHeight, this.getDetailRowHeight());
+        this.rows, this.rowHeight, this.getDetailRowHeight);
     }
   }