+++ src/components/columns/column-cell.directive.ts
@@ -2,5 +2,5 @@ import { Directive, TemplateRef } from '@angular/core';
 
 @Directive({ selector: '[ngx-datatable-cell-template]' })
 export class DataTableColumnCellDirective {
-  constructor(public template: TemplateRef<any>) { };
+  constructor(public template: TemplateRef<any>) { }
 }