+++ src/components/footer/footer.directive.ts
@@ -1,0 +1,19 @@
+import { Input, Output, EventEmitter, Directive, TemplateRef, ContentChild } from '@angular/core';
+import { DataTableFooterTemplateDirective } from './footer-template.directive';
+
+@Directive({ selector: 'ngx-datatable-footer' })
+export class DatatableFooterDirective {
+
+  @Input() footerHeight: number;
+  @Input() totalMessage: string;
+  @Input() selectedMessage: string | boolean;
+  @Input() pagerLeftArrowIcon: string;
+  @Input() pagerRightArrowIcon: string;
+  @Input() pagerPreviousIcon: string;
+  @Input() pagerNextIcon: string;
+
+  @Input()
+  @ContentChild(DataTableFooterTemplateDirective, { read: TemplateRef }) 
+  template: TemplateRef<any>;
+
+}