+++ src/utils/throttle.ts
@@ -9,10 +9,10 @@
  */
 export function throttle(func: any, wait: number, options?: any) {
   options = options || {};
-  let context;
-  let args;
-  let result;
-  let timeout = null;
+  let context: any;
+  let args: any;
+  let result: any;
+  let timeout: any = null;
   let previous = 0;
 
   function later() {
@@ -59,7 +59,7 @@ export function throttle(func: any, wait: number, options?: any) {
  * @returns
  */
 export function throttleable(duration: number, options?: any) {
-  return function innerDecorator(target, key, descriptor) {
+  return function innerDecorator(target: any, key: PropertyKey, descriptor: PropertyDescriptor) {
     return {
       configurable: true,
       enumerable: descriptor.enumerable,